Refactoring the single linked list
